+++ Facade/UserManager.cs
@@ -29,36 +29,19 @@
             return result;
         }
 
-        public async Task<IdentityResult> CreateWithSampleDataAsync(User user, string password, int sampleUserId)
+        public override async Task<IdentityResult> CreateAsync(User user, string password)
         {
-            // await Store.CopySampleDataAsync(sampleUserId, user);
-            
             var result = await base.CreateAsync(user, password);
 
             if (result.Succeeded)
             {
                 await Store.SaveChangesAsync();
 
+                // Send registration alert
                 var emailManager = new EmailManager();
-                await emailManager.SendConfirmationAlert(user);
+                await emailManager.SendRegistrationAlert(user);
             }
             
-            return result;
-        }
-
-        public async Task ResetSampleDataAsync(int userId, int sampleUserId)
-        {
-            await Store.ResetSampleDataAsync(userId, sampleUserId);
-            await Store.SaveChangesAsync();
-        }
-
-        public override async Task<IdentityResult> DeleteAsync(User user)
-        {
-            await Store.DeleteResourcePoolDataAsync(user.Id);
-
-            var result = await base.DeleteAsync(user);
-            await Store.SaveChangesAsync();
-
             return result;
         }
     }